Add adjacencies for kubernetes storage components
Kubernetes storage components such as PV, PVC, StorageClass and
stateful application pod are connected together based on the the
Persistent volume claim name.

Stateful pod and PVC for the same can be connected by means of volume claim
name present in the pod. We are utilizing this to draw adjacency between
stateful pod and it's persistent volume claim.

Further, PVC contains the volume name which is, PV name itself. Hence, we can
show edge for PVC and PV. Similarly we are doing same for storage class.

This will bring higher level visibility for kubernetes storage components.

Signed-off-by: Satyam Zode <satyam.zode@openebs.io>